refactor: refactor data
